# src/main/java/com/personal/trainer/app/model/Exercise.java
package com.personal.trainer.app.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

import java.util.Collection;
import java.util.List;

@Document("exercises")
public class Exercise {

    @Id
    private String id;

    private String name;

    private List<String> muscles;

    private MovementType movementType;

    private ExerciseType exerciseType;

    public Exercise() {}

    public Exercise(String name, List<String> muscles) {
        this.name = name;
        this.muscles = muscles;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Collection<String> getMuscles() {
        return muscles;
    }

    public void setMuscles(List<String> muscles) {
        this.muscles = muscles;
    }

    public MovementType getMovementType() {
        return movementType;
    }

    public void setMovementType(MovementType movementType) {
        this.movementType = movementType;
    }

    public ExerciseType getExerciseType() {
        return exerciseType;
    }

    public void setExerciseType(ExerciseType exerciseType) {
        this.exerciseType = exerciseType;
    }
}